triggers and msg.unparsed
Made it easier to copy a msg obj as long as you have an irc_c

msg.unparsed was set by the triggers subsystem right be fore dispatch.
Instead of modifying msg we copy it first.

@keyword.sub now updates msg.unparsed after making a copy

example bot debug plugin now has '!test sub' which will trigger it
and !test but show how parsing is handled.